+++ CHANGELOG.md
@@ -14,6 +14,7 @@ request to fix it.
 ### Changed
 
 - [lustre/runtime] Fixed a bug where the special-cased handling of "submit" events would apply to unrelated events of the same name dispatched by Web Components.
+- [lustre/runtime] Remove an invalid `@__PURE__` annotation causing warnings in some bundlers and build tools.
 
 ## [v5.7.1] - 2026-14-07